adding more checks
