+++ kernel/include/arch/generic_x86/x86_utils.h
@@ -111,30 +111,6 @@ static ALWAYS_INLINE uptr get_eflags()
 
 extern volatile bool in_panic;
 extern volatile int disable_interrupts_count;
-extern volatile u64 jiffies;
-
-static inline void DEBUG_check_disable_interrupts_count(int int_num)
-{
-#ifdef DEBUG
-
-   int disable_int_c = disable_interrupts_count;
-
-   if (disable_int_c != 0) {
-
-      /*
-       * Disable the interrupts in the lowest-level possible in case,
-       * for any reason, they are actually enabled.
-       */
-      HW_disable_interrupts();
-
-      panic("[generic_interrupt_handler] int_num: %i\n"
-            "disable_interrupts_count: %i (expected: 0)\n"
-            "total system ticks: %llu\n",
-            int_num, disable_int_c, jiffies);
-   }
-
-#endif
-}
 
 
 static inline bool are_interrupts_enabled_int(const char *file, int line)
