+++ kernel/include/irq.h
@@ -3,6 +3,9 @@
 
 #include <common_defs.h>
 
+// TODO: avoiding including this arch-specific header here.
+#include <arch/generic_x86/x86_utils.h>
+
 #define FAULT_DIVISION_BY_ZERO      0
 #define FAULT_DEBUG                 1
 #define FAULT_NMI                   2
@@ -49,9 +52,13 @@ u32 pic_get_imr(void);
 extern volatile int nested_interrupts_count;
 extern volatile int nested_interrupts[32];
 
+void pop_nested_interrupt();
+void push_nested_interrupt(int int_num);
+
 
 static inline int get_curr_interrupt()
 {
+   ASSERT(!are_interrupts_enabled());
    return nested_interrupts[nested_interrupts_count - 1];
 }
 
@@ -63,5 +70,64 @@ static ALWAYS_INLINE bool is_irq(int interrupt_num)
           interrupt_num != SYSCALL_SOFT_INTERRUPT;
 }
 
-void pop_nested_interrupt();
-void push_nested_interrupt(int int_num);
+static inline bool in_interrupt_handler()
+{
+   bool res;
+   disable_interrupts();
+   res = is_irq(get_curr_interrupt());
+   enable_interrupts();
+   return res;
+}
+
+// TODO: re-implement this when SYSENTER is supported as well.
+
+static inline bool in_syscall()
+{
+   bool res = false;
+   disable_interrupts();
+
+   for (int i = nested_interrupts_count - 1; i >= 0; i--) {
+      if (nested_interrupts[i] == 0x80) {
+         res = true;
+         break;
+      }
+   }
+
+   enable_interrupts();
+   return res;
+}
+
+/*
+ * -----------------------------------------------------------------------------
+ *
+ * DEBUG STUFF
+ *
+ * -----------------------------------------------------------------------------
+ */
+
+
+extern volatile int disable_interrupts_count;
+extern volatile u64 jiffies;
+
+static inline void DEBUG_check_disable_interrupts_count(int int_num)
+{
+#ifdef DEBUG
+
+   int disable_int_c = disable_interrupts_count;
+
+   if (disable_int_c != 0) {
+
+      /*
+       * Disable the interrupts in the lowest-level possible in case,
+       * for any reason, they are actually enabled.
+       */
+      HW_disable_interrupts();
+
+      panic("[generic_interrupt_handler] int_num: %i\n"
+            "disable_interrupts_count: %i (expected: 0)\n"
+            "total system ticks: %llu\n",
+            int_num, disable_int_c, jiffies);
+   }
+
+#endif
+}
